Add more detail in the sift section; auto calc pour water weight
Added sift duration and medium grounds weight. Renamed the fields, and
update the form and the view to be little nicer.

Added a ratio input that defaults to 16. If the grounds weight or the
ratio weight is changed, auto update the pour water weight by
multiplying those two and rounding to the nearest integer. This was my
biggest complaint of using the app that I had to switch to my calculator
every time.