Split up pour section of the form
There were just too many inputs in a row, and they weren't really
ordered in the order that I would actually use them. This way I can
enter the water temp and grounds weight and have it auto-calculate the
pour water weight before I start pouring.

Once I start the bloom I can record the bloom water weight that I used
(though I'm guessing this is kind of pointless to record because it's
always going to be approximately the same and not that interesting). And
then I can record the pour and draw down timings and notes on the
technique as I'm pouring and finishing up.